+++ react/components/LoginComponent.js
@@ -1,23 +1,27 @@
-import React, { Component, Fragment } from 'react'
+import React, { Component, Fragment, Suspense } from 'react'
 import {
   Link,
   withRuntimeContext,
   ExtensionPoint,
   useChildBlock,
 } from 'vtex.render-runtime'
-import OutsideClickHandler from 'react-outside-click-handler'
+
 import classNames from 'classnames'
 
 import { IconProfile } from 'vtex.store-icons'
 import Overlay from 'vtex.react-portal/Overlay'
 import { ButtonWithIcon } from 'vtex.styleguide'
 
-import LoginContent from '../LoginContent'
 import { truncateString } from '../utils/format-string'
 import { translate } from '../utils/translate'
 import { LoginPropTypes } from '../propTypes'
 
 import styles from '../styles.css'
+import Loading from './Loading'
+import Popover from './Popover'
+import OutsideClickHandler from './OutsideClickHandler'
+
+const LoginContent = React.lazy(() => import('./LoginContent'))
 
 const ProfileIcon = ({ iconSize, labelClasses, classes }) => {
   const hasIconBlock = Boolean(useChildBlock({ id: 'icon-profile' }))
@@ -55,12 +59,10 @@ class LoginComponent extends Component {
       onProfileIconClick,
       sessionProfile,
       showIconProfile,
-      runtime: {
-        history: {
-          location: { pathname },
-        },
-      },
+      runtime: { history },
     } = this.props
+
+    const pathname = history && history.location && history.location.pathname
 
     const iconClasses = 'flex items-center'
     const iconLabel = iconLabelProfile || translate('store/login.signIn', intl)
@@ -145,24 +147,16 @@ class LoginComponent extends Component {
           {isBoxOpen && (
             <Overlay>
               <OutsideClickHandler onOutsideClick={onOutSideBoxClick}>
-                <div
-                  className={`${styles.box} z-max absolute`}
-                  style={mirrorTooltipToRight ? { left: 50 } : { right: -50 }}
-                >
-                  <div
-                    className={`${styles.arrowUp} absolute top-0 ${
-                      mirrorTooltipToRight ? 'left-0 ml3' : 'right-0 mr3'
-                    } shadow-3 bg-base rotate-45 h2 w2`}
-                  />
-                  <div className={`${styles.contentContainer} shadow-3 mt3`}>
+                <Popover mirrorTooltipToRight={mirrorTooltipToRight}>
+                  <Suspense fallback={<Loading />}>
                     <LoginContent
                       profile={sessionProfile}
                       loginCallback={this.onClickLoginButton}
                       isInitialScreenOptionOnly
                       {...this.props}
                     />
-                  </div>
-                </div>
+                  </Suspense>
+                </Popover>
               </OutsideClickHandler>
             </Overlay>
           )}